requirements.txt
aiogram>=3.10.0
gigachat>=0.1.30
python-dotenv>=1.0.0